pyproject.toml
[tool.poetry]
name = "resume-rag"
version = "0.1.0"
description = "Local RAG résumé-tailoring API"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = { extras = ["standard"], version = "^0.29" }
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
pgvector = "^0.2"
sentence-transformers = "^2.6"
torch = "^2.2"
python-dotenv = "^1.0"
python-multipart = "^0.0.9"
pypdf2 = "^3.0"
openai = "^1.12"
requests = "^2.31"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.4"
mypy = "^1.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
